app/requirements.txt
dotenv-python==0.0.1
docling==2.15.1
chromadb==0.6.2
sentence-transformers==3.3.1
pydantic==2.10.5
llama-index==0.12.10
llama-index-vector-stores-chroma==0.4.1
llama-index-embeddings-huggingface==0.5.0
llama-index-readers-docling==0.3.2
llama-index-node-parser-docling==0.3.0
llama-index-llms-huggingface==0.4.2
llama-index-llms-huggingface-api==0.3.1
llama-index-llms-groq==0.3.1
llama-index-readers-file==0.4.3
icecream==2.1.4